Initial implementation of a Hibernate Listener
This change adds the `BackboneEventListener` class that can be
registered with Hibernate as an event listener on
`post-collection-update` and `post-collection-remove` events. It will
disect the entities affected by the Hibernate event and place an array
of JSON objects with their entity types and IDs onto the Hazelcast bus
under the `ome.model.change` address.